+++ src/core/types.ts
@@ -94,6 +94,7 @@ export interface IndexedDatabaseTransaction {
   onerror: ((event: unknown) => void) | null;
   onabort: ((event: unknown) => void) | null;
   readonly error: Error | null;
+  readonly mode: IndexedDatabaseTransactionMode;
   objectStore(name: string): IndexedDatabaseObjectStore;
   abort(): void;
 }